% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{avar}
\alias{avar}
\title{Calculate the asymptotic variance.}
\usage{
avar(x, sigma, type, parallel)
}
\arguments{
\item{x}{Data to estimate alpha on.}

\item{sigma}{Covariance matrix of x.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}.}

\item{parallel}{If \code{TRUE}, makes calculations under the assumption of a
parallel model.}

\item{transform}{Choose between \code{"fisher"} and \code{"none"}.}
}
\value{
Asymptotic variance.
}
\description{
Calculate the asymptotic variance.
}
\keyword{internal}
