% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{avar_elliptical}
\alias{avar_elliptical}
\title{Theoretical variance for the elliptical model.}
\usage{
avar_elliptical(x, sigma, type)
}
\arguments{
\item{x}{Data matrix.}

\item{sigma}{Covariance matrix.}

\item{type}{Either \code{"normal"} or \code{"elliptical"}.}
}
\value{
Theoretical variance.
}
\description{
Theoretical variance for the elliptical model.
}
\keyword{internal}
