% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{avar_parallel_adf}
\alias{avar_parallel_adf}
\title{Variance under parallel and type = \code{adf}.}
\usage{
avar_parallel_adf(alpha, sigma, gamma)
}
\description{
Variance under parallel and type = \code{adf}.
}
