% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{avar_parallel_elliptical}
\alias{avar_parallel_elliptical}
\title{Variance under parallel elliptical model.}
\usage{
avar_parallel_elliptical(alpha, x, type)
}
\arguments{
\item{alpha}{Coefficient alpha or standardized alpha.}

\item{x}{D}
}
\description{
Variance under parallel elliptical model.
}
