% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{covmat}
\alias{covmat}
\title{Transform lambda and sigma to a covariance matrix.}
\usage{
covmat(lambda, sigma)
}
\arguments{
\item{lambda}{Vector of loadings.}

\item{sigma}{Vector of standard deviations.}
}
\value{
The covariance matrix implied by lambda and sigma.
}
\description{
Transform lambda and sigma to a covariance matrix.
}
\keyword{internal}
