% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{gamma_mat_normal}
\alias{gamma_mat_normal}
\title{Calculate gamma matrix under normality.}
\usage{
gamma_mat_normal(sigma)
}
\arguments{
\item{sigma}{Covariance matrix of the data.}
}
\value{
Gamma matrix under normality.
}
\description{
This is also the "uncorrected" gamma matrix under ellipticity.
}
\keyword{internal}
