% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{gs}
\alias{gs}
\title{The gs vector used in the asymptotic variance of standardized alpha.}
\usage{
gs(phi)
}
\arguments{
\item{phi}{Correlation matrix.}
}
\value{
The gs vector.
}
\description{
The gs vector used in the asymptotic variance of standardized alpha.
}
\keyword{internal}
