% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{kurtosis}
\alias{kurtosis}
\title{Calculate unbiased sample kurtosis.}
\usage{
kurtosis(x)
}
\arguments{
\item{x}{Matrix of valus.}
}
\value{
Unbiased sample kurtosis.
}
\description{
Calculate unbiased sample kurtosis.
}
\keyword{internal}
