% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{kurtosis_correction}
\alias{kurtosis_correction}
\title{Calculate kurtosis correction}
\usage{
kurtosis_correction(x, type)
}
\arguments{
\item{x}{Matrix of values}

\item{type}{The type of correction, either "normal" or "elliptical".}
}
\description{
Calculate kurtosis correction
}
\keyword{internal}
