% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{limits}
\alias{limits}
\title{Calculate limits of a confidence interval.}
\usage{
limits(alternative, conf_level)
}
\arguments{
\item{alternative}{Alternative choosen.}

\item{conf_level}{Confidence level.}
}
\description{
Calculate limits of a confidence interval.
}
\keyword{internal}
