% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{psi_mat}
\alias{psi_mat}
\title{Psi matrix}
\usage{
psi_mat(x, sigma, type = "adf")
}
\arguments{
\item{x}{A numeric matrix of observations.}

\item{sigma}{Covariance matrix.}

\item{type}{One of \code{adf}, \code{normal} and \code{elliptical}.}
}
\value{
The sample estimate of the psi matrix.
}
\description{
Calculate the psi matrix from a matrix of observations.
}
\keyword{internal}
