% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilities.R
\name{reliability}
\alias{reliability}
\alias{alpha_bias}
\alias{omega}
\alias{omega_std}
\alias{alpha}
\alias{alpha_std}
\title{Reliability coefficients}
\usage{
alpha_bias(sigma, lambda, w = rep(1, length(lambda)))

omega(sigma, lambda)

omega_std(sigma, lambda)

alpha(sigma, lambda)

alpha_std(sigma, lambda)
}
\arguments{
\item{sigma}{For \code{alpha} and \code{alpha_std}, either apositive definite
covariance matrix or a vector of standard deviations. A vector of
standard deviations for \code{omega} and \code{omega_std}.}

\item{lambda}{Vector of loadings.}
}
\value{
The congeneric reliability standardized reliability,
coefficient alpha, standardized, orsigma coefficient alpha.
}
\description{
The congeneric reliability and standardized reliability; also the
bias \code{b} so that \code{omega} - \code{alpha} = \code{b}.
}
\keyword{internal}
