% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{studentized_boots}
\alias{studentized_boots}
\title{Studentized bootstrap estimates using transformers.}
\usage{
studentized_boots(n_reps, x, type, parallel, transformer, standardized = FALSE)
}
\arguments{
\item{n_reps}{Number of bootstrap repetitions.}

\item{x}{Data to estimate alpha on.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}.}

\item{parallel}{If \code{TRUE}, makes calculations under the assumption of a
parallel model.}

\item{transformer}{A \code{transformer} object.}

\item{standardized}{If \code{TRUE}, calculates the standardized alpha. Calculates
coefficient alpha otherwise.}
}
\value{
Studentized bootstrap estimates.
}
\description{
Studentized bootstrap estimates using transformers.
}
\keyword{internal}
