% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{var_ell}
\alias{var_ell}
\title{Theoretical variance for the elliptical model.}
\usage{
var_ell(sigma, kurt = 0, parallel = FALSE)
}
\arguments{
\item{sigma}{Covariance matrix.}

\item{kurt}{Common kurtosis parameter.}

\item{parallel}{If \code{TRUE}, calculates the variance under assumption of
a parallel model.}
}
\value{
Theoretical variance.
}
\description{
Theoretical variance for the elliptical model.
}
\keyword{internal}
