% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{alphaci}
\alias{alphaci}
\alias{alphaci_std}
\title{Estimate coefficient alpha with a confidence interval.}
\usage{
alphaci(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  parallel = FALSE,
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)

alphaci_std(
  x,
  type = c("adf", "elliptical", "normal"),
  transform = "none",
  parallel = FALSE,
  conf_level = 0.95,
  alternative = c("two.sided", "greater", "less"),
  bootstrap = FALSE,
  n_reps = 1000
)
}
\arguments{
\item{x}{Data to estimate alpha on.}

\item{type}{Type of confidence interval. Either \code{adf}, \code{elliptical}, or
\code{normal}.}

\item{transform}{One of \code{"none"}, \code{"log"}, \code{"fisher"}, and \verb{"arcsin}.
Defaults to \code{"none"}.}

\item{parallel}{If \code{TRUE}, makes calculations under the assumption of a
parallel model. Default to \code{FALSE}.}

\item{conf_level}{Confidence level.}

\item{alternative}{Direction of the confidence interval.}

\item{bootstrap}{If \code{TRUE}, performs a studentized bootstrap with \code{n_reps}
repetitions. Defaults to \code{FALSE}.}

\item{n_reps}{Number of bootstrap samples if \code{bootstrap = TRUE}. Ignored if
\code{bootstrap = FALSE}. Defaults to \code{1000}.}
}
\value{
An appropriate object.
}
\description{
This function makes use of \code{future.apply} when bootstrapping. Supports
both ordinary coefficient alpha (\code{alphaci}) and standardized coefficient
alpha (\code{alphaci_std}).
}
