% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{gamma_mat}
\alias{gamma_mat}
\title{Gamma matrix}
\usage{
gamma_mat(x)
}
\arguments{
\item{x}{A numeric matrix of observations.}
}
\value{
The sample estimate of the gamma matrix.
}
\description{
Calculate the gamma matrix from a matrix of observations.
}
\keyword{internal}
