% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{gs}
\alias{gs}
\title{The gs vector used in the asymptotic variance of standardized alpha.}
\usage{
gs(phi)
}
\arguments{
\item{x}{Correlation matrix.}

\item{parallel}{If \code{TRUE}, returns the value of \code{gs} under the parallel
model.}
}
\value{
The gs vector.
}
\description{
The gs vector used in the asymptotic variance of standardized alpha.
}
\keyword{internal}
